use: switch form PATH to fish_user_paths
This might impact a few people, but I think this approach is correct as
it prepends, but also gives you flexibility to manipulate this
prepending part if you don't want ruby path to be very first amongst
these.

An upcoming idea is to allow users to configure if it should be
prepended or appended to the `fish_user_paths`. And also put in some
bundler support with binstubs.

Basically integrate this:

https://github.com/terlar/dotfiles/tree/f6a132761a25a5ccda429dafbf0fa1dc17376797/config/fish/plugins/bundler