config: add support for prepend path config
You can now choose if you want to prepend or append the ruby when you
switch. Please note that it is prepending or appending to the
`fish_user_paths` which means it will be before anything else in the
regular path even if you append. This is just to allow you control your
own user added paths and the order of them.