fix rouge functions output
This were two errors, first the fry_installer_toggle should only be
triggered by the change of variable `fry_installer`. And it needs to
initialize first time, so the not set logic that detects what installers
you have installed needs to go after the on variable triggered function.

In the test-helper we referred to a bad installer, installers should be
prefixed with `fry-installer-`

Also unsetting the variables will mess with the environment, so better
let them be until script finishes.