restructure Makefile and installation method
This flattens the project structure and moves the completions, functions
and init script to their own directories under the project root level.

The installation will now use the new available vendor directories
provided by fish. This makes it so that users don't have to add this to
their bootstrap scripts manually etc.